server/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Firebase Admin SDK
firebase-admin==6.2.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# HTTP client for external APIs
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# AI/ML libraries
openai==1.3.5
scikit-learn==1.3.2

# Financial data APIs
yfinance==0.2.28
alpha-vantage==2.3.1

# Environment management
python-dotenv==1.0.0

# Security
cryptography==41.0.7
bcrypt==4.1.1

# Task queue (for background processing)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0